+++ store/badger/badger_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/dfuse-io/kvdb/store"
 	"github.com/dfuse-io/kvdb/store/storetest"
+	"github.com/dfuse-io/logging"
 	"github.com/stretchr/testify/require"
 )
 
@@ -21,13 +22,14 @@ func TestAll(t *testing.T) {
 }
 
 func NewTestBadgerFactory(t *testing.T, testDBFilename string) storetest.DriverFactory {
-	return func(opts ...store.Option) (store.KVStore, storetest.DriverCleanupFunc) {
+	return func(opts ...store.Option) (store.KVStore, *storetest.DriverCapabilities, storetest.DriverCleanupFunc) {
 		dir, err := ioutil.TempDir("", "kvdb-badger")
 		require.NoError(t, err)
 		dsn := fmt.Sprintf("badger://%s", path.Join(dir, testDBFilename))
 		kvStore, err := store.New(dsn, opts...)
 		require.NoError(t, err)
-		return kvStore, func() {
+
+		return kvStore, storetest.NewDriverCapabilities(), func() {
 			err := os.RemoveAll(dir)
 			require.NoError(t, err)
 		}