+++ store/netkv/options.go
@@ -1,4 +1,5 @@
 package netkv
 
 func (s *Store) EnableEmpty() {
+	zlog.Info("discarding possible empty value on store implementation, not required for this store")
 }