+++ store/storetest/testing.go
@@ -13,9 +13,17 @@ func init() {
 	debug = os.Getenv("DEBUG") != ""
 }
 
+type DriverCapabilities struct {
+	SupportsEmptyValue bool
+}
+
+func NewDriverCapabilities() *DriverCapabilities {
+	return &DriverCapabilities{SupportsEmptyValue: true}
+}
+
 type DriverCleanupFunc func()
-type DriverFactory func(opts ...store.Option) (store.KVStore, DriverCleanupFunc)
+type DriverFactory func(opts ...store.Option) (store.KVStore, *DriverCapabilities, DriverCleanupFunc)
 
 func TestAll(t *testing.T, driverName string, driverFactory DriverFactory) {
-	TestAllKVStore(t, driverName, driverFactory)
+	testAllKVStore(t, driverName, driverFactory)
 }