+++ store/tikv/options.go
@@ -1,5 +1,6 @@
 package tikv
 
 func (s *Store) EnableEmpty() {
+	zlog.Info("enabling possible empty value on store implementation")
 	s.emptyValuePossible = true
 }